feat: partial frontend rewrite
