+++ web/vite.config.ts
@@ -2,5 +2,15 @@ import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
 
 export default defineConfig({
-	plugins: [sveltekit()]
+	plugins: [sveltekit()],
+	resolve: {
+		alias: {
+			$static: '/',
+		}
+	},
+	server: {
+		fs: {
+			strict: false
+		}
+	}
 });
